Docker/Aggregation/Enrichment/requirements.txt
filelock==3.15.4
nats-py==2.7.2
ndjson==0.3.1
python-dotenv==0.21.1
redis==5.0.4
requests>2.32.0
tldextract==5.1.1
validators==0.22.0
watchdog==6.0.0